perf: reduce hotspot test startup and timeout costs
